Client account and Pooling

+++ bankease/src/view/ClientsList.java
@@ -2,11 +2,17 @@ package view;
 
 import javax.swing.JFrame;
 import javax.swing.JPanel;
+
+import controller.AccountListController;
 import controller.ClientHandler;
+import model.Account;
 import model.Client;
 import javax.swing.JLabel;
+
+import java.awt.Color;
 import java.awt.Font;
 import javax.swing.JList;
+import javax.swing.JOptionPane;
 import javax.swing.JScrollPane;
 import javax.swing.event.ListSelectionListener;
 import javax.swing.event.ListSelectionEvent;
@@ -32,14 +38,12 @@ public class ClientsList extends JFrame {
 		panel.add(lblNewLabel);
 		JScrollPane scrollPane = new JScrollPane();
 
-		scrollPane.setBounds(100, 60, 460, 400);
+		scrollPane.setBounds(100, 60, 460, 300);
 		getContentPane().add(scrollPane);
 
 		Client[] clientsList = ClientHandler.listOfClients();
 		JList<Client> list = new JList<Client>();
 		list.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		//list.setFixedCellHeight(44);
-		//list.setFont(list.getFont().deriveFont(16.0f));
 		list.setListData(clientsList);
 
 		scrollPane.setViewportView(list);
@@ -47,16 +51,6 @@ public class ClientsList extends JFrame {
 		JPanel panel_1 = new JPanel();
 		panel_1.setBounds(0, 540, 660, 39);
 		getContentPane().add(panel_1);
-
-		JButton btnDetailsClient = new JButton("Détails Client");
-		btnDetailsClient.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-			}
-		});
-
-		btnDetailsClient.setBounds(0, 0, 100, 30);
-		btnDetailsClient.setEnabled(false);
-		panel_1.add(btnDetailsClient);
 
 		JButton btnModifyClient = new JButton("Modifier Client");
 		btnModifyClient.addActionListener(new ActionListener() {
@@ -77,11 +71,15 @@ public class ClientsList extends JFrame {
 		btnDeleteClient.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				Client selectedClient = list.getSelectedValue();
-				if (selectedClient != null) {
-					ClientHandler.deleteClient(selectedClient.getClientId());
-					Client[] clientsList = ClientHandler.listOfClients();
-					list.setListData(clientsList);
-					scrollPane.setViewportView(list);
+				int result = JOptionPane.showConfirmDialog(ClientsList.this, "Retirer " + selectedClient.getClientDescription() + " des clients ?", "Confirmation avant suppression", JOptionPane.YES_NO_OPTION);
+				if (result == JOptionPane.YES_OPTION) {
+					if (selectedClient != null) {
+						ClientHandler.deleteClient(selectedClient.getClientId());
+						Client[] clientsList = ClientHandler.listOfClients();
+						list.setListData(clientsList);
+						scrollPane.setViewportView(list);
+						selectedClient = null;
+						}
 				}
 			}
 		});
@@ -91,13 +89,12 @@ public class ClientsList extends JFrame {
 		panel_1.add(btnDeleteClient);
 
 		JButton btnAddClient = new JButton("Ajouter Client");
-		btnDetailsClient.setSize(100, 30);
 		panel_1.add(btnAddClient);
-		
+
 		JPanel panel_2 = new JPanel();
 		panel_2.setBounds(0, 618, 660, 48);
 		getContentPane().add(panel_2);
-		
+
 		JButton btnAccounts = new JButton("Voir comptes");
 		btnAccounts.setEnabled(false);
 
@@ -111,6 +108,47 @@ public class ClientsList extends JFrame {
 		});
 		panel_2.add(btnAccounts);
 
+		JPanel panel_3 = new JPanel();
+		panel_3.setBounds(100, 370, 460, 143);
+		getContentPane().add(panel_3);
+		panel_3.setLayout(null);
+
+		JLabel clientBirth = new JLabel("New label");
+		clientBirth.setFont(new Font("Tahoma", Font.PLAIN, 14));
+		clientBirth.setBounds(220, 10, 230, 24);
+		panel_3.add(clientBirth);
+		clientBirth.setVisible(false);
+
+		JLabel clientAddress = new JLabel("New label");
+		clientAddress.setFont(new Font("Tahoma", Font.PLAIN, 14));
+		clientAddress.setBounds(220, 44, 230, 24);
+		panel_3.add(clientAddress);
+		clientAddress.setVisible(false);
+
+		JLabel clientPhone = new JLabel("New label");
+		clientPhone.setFont(new Font("Tahoma", Font.PLAIN, 14));
+		clientPhone.setBounds(220, 78, 230, 24);
+		panel_3.add(clientPhone);
+		clientPhone.setVisible(false);
+
+		JLabel clientBirthLabel = new JLabel("Date de naissance :");
+		clientBirthLabel.setFont(new Font("Tahoma", Font.PLAIN, 14));
+		clientBirthLabel.setBounds(10, 10, 186, 24);
+		panel_3.add(clientBirthLabel);
+		clientBirthLabel.setVisible(false);
+
+		JLabel clientAddressLabel = new JLabel("Adresse :");
+		clientAddressLabel.setFont(new Font("Tahoma", Font.PLAIN, 14));
+		clientAddressLabel.setBounds(10, 44, 186, 24);
+		panel_3.add(clientAddressLabel);
+		clientAddressLabel.setVisible(false);
+
+		JLabel clientPhoneLabel = new JLabel("Téléphone :");
+		clientPhoneLabel.setFont(new Font("Tahoma", Font.PLAIN, 14));
+		clientPhoneLabel.setBounds(10, 78, 186, 24);
+		panel_3.add(clientPhoneLabel);
+		clientPhoneLabel.setVisible(false);
+
 		btnAddClient.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				setVisible(false);
@@ -123,10 +161,32 @@ public class ClientsList extends JFrame {
 			@Override
 			public void valueChanged(ListSelectionEvent client) {
 				if (!client.getValueIsAdjusting()) {
-					btnDetailsClient.setEnabled(true);
-					btnModifyClient.setEnabled(true);
-					btnDeleteClient.setEnabled(true);
-					btnAccounts.setEnabled(true);
+					Client selectedClient = list.getSelectedValue();
+					if (selectedClient != null) {
+						btnModifyClient.setEnabled(true);
+						btnDeleteClient.setEnabled(true);
+						btnAccounts.setEnabled(true);
+						clientBirthLabel.setVisible(true);
+						clientBirth.setVisible(true);
+						clientBirth.setText(selectedClient.getClientBirthdate().toString());
+						clientAddressLabel.setVisible(true);
+						clientAddress.setVisible(true);
+						clientAddress.setText(selectedClient.getClientAddress());
+						clientPhoneLabel.setVisible(true);
+						clientPhone.setVisible(true);
+						clientPhone.setText(selectedClient.getClientPhone().toString());
+					} else {
+						btnModifyClient.setEnabled(false);
+						btnDeleteClient.setEnabled(false);
+						btnAccounts.setEnabled(false);
+						clientBirthLabel.setVisible(false);
+						clientBirth.setVisible(false);
+						clientAddressLabel.setVisible(false);
+						clientAddress.setVisible(false);
+						clientPhoneLabel.setVisible(false);
+						clientPhone.setVisible(false);						
+					}
+
 				}
 			}
 		});
